Embeddings stored as floats

+++ cpp/include/cuml/manifold/umap.hpp
@@ -23,21 +23,21 @@
 namespace ML {
 
 void transform(const cumlHandle &handle, float *X, int n, int d, float *orig_X,
-               int orig_n, double *embedding, int embedding_n,
-               UMAPParams *params, double *transformed);
+               int orig_n, float *embedding, int embedding_n,
+               UMAPParams *params, float *transformed);
 
 void find_ab(const cumlHandle &handle, UMAPParams *params);
 
 void fit(const cumlHandle &handle,
          float *X,  // input matrix
          float *y,  // labels
-         int n, int d, UMAPParams *params, double *embeddings);
+         int n, int d, UMAPParams *params, float *embeddings);
 
 void fit(const cumlHandle &handle,
          float *X,  // input matrix
          int n,     // rows
          int d,     // cols
-         UMAPParams *params, double *embeddings);
+         UMAPParams *params, float *embeddings);
 
 class UMAP_API {
   float *orig_X;
@@ -60,7 +60,7 @@ class UMAP_API {
    * @param embeddings
    *        an array to return the output embeddings of size (n_samples, n_components)
    */
-  void fit(float *X, int n, int d, double *embeddings);
+  void fit(float *X, int n, int d, float *embeddings);
 
   /**
    * Fits a supervised UMAP model
@@ -75,7 +75,7 @@ class UMAP_API {
    * @param embeddings
    *        an array to return the output embeddings of size (n_samples, n_components)
    */
-  void fit(float *X, float *y, int n, int d, double *embeddings);
+  void fit(float *X, float *y, int n, int d, float *embeddings);
 
   /**
    * Project a set of X vectors into the embedding space.
@@ -92,8 +92,8 @@ class UMAP_API {
    * @param out
    *        pointer to array for storing output embeddings (n, n_components)
    */
-  void transform(float *X, int n, int d, double *embedding, int embedding_n,
-                 double *out);
+  void transform(float *X, int n, int d, float *embedding, int embedding_n,
+                 float *out);
 
   /**
    * Get the UMAPParams instance
